refactor: js파일 구조 변경 - utility.js 추가 - signin, signup 페이지 공통 스크립트를 utility.js로 이동 	new file:   js/utility.js 	modified:   signin/index.html 	modified:   signin/index.js
