+++ signin/index.html
@@ -58,6 +58,7 @@
         <a href="https://www.kakaocorp.com/page/"><img src="../public/icon/Kakao.svg" /></a>
       </div>
     </div>
-    <script src="./index.js"></script>
+    <script type="module" src="./index.js"></script>
+    <script type="module" src="../js/utility.js"></script>
   </body>
 </html>
